+++ src/ocean_emulators/utils/train.py
@@ -5,8 +5,8 @@ from pathlib import Path
 import torch
 from xarray_einstats.einops import rearrange  # noqa: F401
 
-from ocean_emulators.datasets import InferenceDataset
-from ocean_emulators.utils.data import LoadStats, TrainData
+from ocean_emulators.datasets import InferenceDataset, TrainData
+from ocean_emulators.utils.data import LoadStats
 
 
 def pairwise(iterable):
